Rename `CommandAssignee` to `Assignee`
