+++ server/src/main/java/io/spine/server/command/Assignee.java
@@ -26,30 +26,13 @@
 
 package io.spine.server.command;
 
-import io.spine.server.command.model.CommandAssigneeSignature;
-import io.spine.server.command.model.given.handler.InvalidAssignee;
-import io.spine.server.command.model.given.handler.ValidAssignee;
-import io.spine.server.model.MethodSignatureTest;
-import org.junit.jupiter.api.DisplayName;
-
-import java.lang.reflect.Method;
-import java.util.stream.Stream;
-
-@DisplayName("`CommandAssigneeSignature` should")
-class CommandAssigneeSignatureTest extends MethodSignatureTest<CommandAssigneeSignature> {
-
-    @Override
-    protected Stream<Method> validMethods() {
-        return methodsAnnotatedWith(Assign.class, ValidAssignee.class).stream();
-    }
-
-    @Override
-    protected Stream<Method> invalidMethods() {
-        return methodsAnnotatedWith(Assign.class, InvalidAssignee.class).stream();
-    }
-
-    @Override
-    protected CommandAssigneeSignature signature() {
-        return new CommandAssigneeSignature();
-    }
+/**
+ * Declares one or more methods, each {@linkplain Assign assigned} to handle commands
+ * of a specific type.
+ *
+ * <p>A handling method may reject the received command or produce events in response.
+ *
+ * @see Assign @Assign
+ */
+public interface Assignee extends CommandReceiver {
 }
